Move models and routes folders into lib
