+++ lib/packages.js
@@ -1,10 +1,10 @@
-var PackageModel = require('../models/package').PackageModel
+var PackageModel = require('../lib/models/package').PackageModel
 	, error = require('./error')
 	, users = require('./users')
 	, async = require('async')
 	, _ = require('underscore')
 	, mongoose = require('mongoose')
-	, UserModel = require('../models/user').UserModel
+	, UserModel = require('../lib/models/user').UserModel
     , fs = require('fs')
 	, crypto = require('crypto')
     , storage = require('./storage');
