+++ lib/routes/index.js
@@ -1,0 +1,10 @@
+
+/*
+ * GET home page.
+ */
+
+exports.index = function(req, res){
+
+  res.render('index', { title: 'greg' });
+
+};
