+++ lib/routes/stats.js
@@ -1,6 +1,6 @@
-var error = require('../lib/error')
+var error = require('../error')
   , mongoose = require('mongoose')
-  , stats = require('../lib/stats')
+  , stats = require('../stats')
 
 exports.DEFAULT_LIMIT = 6;
