+++ lib/users.js
@@ -1,5 +1,5 @@
-var UserModel = require('../models/user').UserModel
-	, PackageModel = require('../models/package').PackageModel
+var UserModel = require('../lib/models/user').UserModel
+	, PackageModel = require('../lib/models/package').PackageModel
 	, error = require('./error')
 	, _ = require('underscore')
 	, async = require('async');
@@ -12,16 +12,16 @@ var UserModel = require('../models/user').UserModel
  * @api private
  */
 
+var debugUserName = 'test';
+
 exports.initDebugUser = function() {
 
-	var username = 'test';
-
-	UserModel.findOne({username: username}, function(err, user){
+	UserModel.findOne({username: debugUserName}, function(err, user){
 
 	  if (err || !user) {
 
 	    console.log('Attempting to create new debug user...');
-	    var master_user = new UserModel({username: username});
+	    var master_user = new UserModel({username: debugUserName});
 
 	    master_user.save(function(err) {
 	    	if (err) {
@@ -30,12 +30,13 @@ exports.initDebugUser = function() {
 	    	console.log('Successfully created new debug user...')
 	    });
 	  }
-	  else {
-	  	console.log('DEBUG USER: ', user);
-	  }
 
 	});
 
+}
+
+exports.cleanupDebugUser = function(){
+    UserModel.FindOne
 }
 
 /**
