--- routes/index.js
@@ -1,10 +1,0 @@
-
-/*
- * GET home page.
- */
-
-exports.index = function(req, res){
-
-  res.render('index', { title: 'greg' });
-
-};
